Just one templatized ReferenceCounted class.
