+++ CesiumRasterOverlays/include/CesiumRasterOverlays/RasterOverlay.h
@@ -6,7 +6,7 @@
 #include <CesiumAsync/IAssetAccessor.h>
 #include <CesiumGltf/Ktx2TranscodeTargets.h>
 #include <CesiumUtility/IntrusivePointer.h>
-#include <CesiumUtility/ReferenceCountedNonThreadSafe.h>
+#include <CesiumUtility/ReferenceCounted.h>
 
 #include <nonstd/expected.hpp>
 #include <spdlog/fwd.h>
